moved the breeder related functions from the pedigree app into the breeder app
